Moved cmd module to driver/cmd
