+++ adapters/driver/cmd/rabbitmq.go
