+++ adapters/driver/cmd/server.go
@@ -3,12 +3,12 @@ package cmd
 import (
 	"context"
 	"errors"
-	"github.com/NicklasWallgren/go-template/adapters/driver/api/common"
-	"github.com/NicklasWallgren/go-template/adapters/driver/api/middlewares"
-	"github.com/NicklasWallgren/go-template/adapters/driver/api/routes"
 	"log"
 	"net/http"
 	"time"
+	"github.com/NicklasWallgren/go-template/adapters/driver/api/common"
+	"github.com/NicklasWallgren/go-template/adapters/driver/api/middlewares"
+	"github.com/NicklasWallgren/go-template/adapters/driver/api/routes"
 
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
